+++ src/MoreLeaderboards/SelectRegion.cpp
@@ -24,10 +24,10 @@ std::vector<std::string> SelectRegion::getWords(std::string s, std::string delim
     return res;
 }
 
-void SelectRegion::scene() {
-    auto popup = new SelectRegion();
+void SelectRegion::scene(const std::function<void(int)>& callback) {
+    auto popup = new SelectRegion(callback);
 
-    if (popup && popup->setup()) {
+    if (popup && popup->init(250.0f, 210.0f)) {
         popup->autorelease();
         CCDirector::sharedDirector()->getRunningScene()->addChild(popup);
     } else {
@@ -53,9 +53,9 @@ void SelectRegion::fadeLoadingCircle() {
 };
 
 void SelectRegion::startLoading() {
-    /*loading_circle = LoadingCircle::create();
+    loading_circle = LoadingCircle::create();
     loading_circle->setParentLayer(this);
-    loading_circle->show();*/
+    loading_circle->show();
 
     geode::utils::web::WebRequest request = web::WebRequest();
     const geode::utils::MiniFunction<void(std::string const&)> then = [this](std::string const& data) {
@@ -87,26 +87,22 @@ void SelectRegion::loadPage() {
     CCArray* data = CCArray::create();
 
     for (std::string const& mod : displayedData) {
-        auto cell = SelectRegionCell::create(mod);
+        auto cell = SelectRegionCell::create(mod, [this](int id) {
+            selectedRegion(id);
+            removeFromParent();
+        });
         data->addObject(cell);
     }
 
-    auto listview = ListView::create(data, 20, 200, 180);
-    m_mainLayer->addChild(listview);
+    auto border = OldBorder::create(ListView::create(data, 30, 218, 178), { 0xBF, 0x72, 0x3E, 0xFF }, {220, 180}, {1, 1});
+    border->setPosition(m_mainLayer->getContentSize() / 2 - border->getContentSize() / 2);
+    setTouchPriority(100);
+    m_mainLayer->addChild(border);
 }
 
 void SelectRegion::handle_request(std::string const& data) {
-    log::debug("mod: {}", data);
-
     if (data != "-1") {
         displayedData = getWords(data, "|");
-
-        /*while (mods.size() > 0) {
-            std::string mod = mods[0];
-            log::debug("mod: {}", mod);
-            displayedData.push_back(mod);
-            mods.erase(mods.begin());
-        };*/
+        loadPage();
     }
-    loadPage();
 }