+++ changelog.md
@@ -1,6 +1,7 @@
 # v1.3.0
     - Port to Geometry Dash 2.206
     - Increased leaderboards top to INFINITE
+    - Select leaderboards by countries
     - Daily chest notifications
     - Epic Gauntlets redesign
     - Added previews for locked icons with colors